refactor: split the budget history chart into its parts
